create scalability report directory
